Moved new Windows header to ola/win folder
