+++ common/base/Init.cpp
@@ -43,7 +43,7 @@
 #include <stdio.h>
 #define WIN32_LEAN_AND_MEAN
 #define VC_EXTRALEAN
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #else
 #include <sys/resource.h>
 #endif
