+++ common/base/Logging.cpp
@@ -31,7 +31,7 @@
 
 #ifdef _WIN32
 #define VC_EXTRALEAN
-#include <ola/CleanWindows.h>
+#include <ola/win/CleanWindows.h>
 #include <io.h>
 #else
 #include <syslog.h>
