+++ common/http/HTTPServer.cpp
@@ -28,7 +28,7 @@
 #include <ola/web/JsonWriter.h>
 
 #ifdef _WIN32
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #endif
 
 #include <fstream>
