+++ common/io/SelectServer.cpp
@@ -21,7 +21,7 @@
 #include "ola/io/SelectServer.h"
 
 #ifdef _WIN32
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #else
 #include <sys/select.h>
 #endif
