+++ common/io/SelectServerTest.cpp
@@ -24,7 +24,7 @@
  */
 
 #ifdef _WIN32
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #endif
 
 #include <cppunit/extensions/HelperMacros.h>
