+++ common/network/TCPSocket.cpp
@@ -34,7 +34,7 @@
 #include <unistd.h>
 
 #ifdef _WIN32
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #include <Ws2tcpip.h>
 #include <winioctl.h>
 #else
