+++ include/ola/http/HTTPServer.h
@@ -34,7 +34,7 @@
 #include <stdlib.h>
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #else
 #include <sys/select.h>
 #include <sys/socket.h>
