+++ include/ola/win/CleanWinSock2.h
@@ -15,8 +15,7 @@
  *
  * CleanWinSock2.h
  * A common header that removes many of the name-space clutter that windows.h
- * and WinSock2.h
- * creates
+ * and WinSock2.h creates
  * Copyright (C) 2014 Sean Sill
  */
