+++ ola/AutoStart.cpp
@@ -23,7 +23,7 @@
 #include <string.h>
 #ifdef _WIN32
 #define VC_EXTRALEAN
-#include <ola/CleanWindows.h>
+#include <ola/win/CleanWindows.h>
 #include <tchar.h>
 #else
 #include <sys/wait.h>
