+++ plugins/osc/OSCNode.cpp
@@ -19,7 +19,7 @@
  */
 
 #ifdef _WIN32
-#include <ola/CleanWinSock2.h>
+#include <ola/win/CleanWinSock2.h>
 #endif
 
 #include <ola/Callback.h>
